// src/_proto/Protos/billing.ts
export interface AddPaymentMethodRequest {
    /**
     * @generated from protobuf oneof: paymentMethod
     */
    paymentMethod: {
        oneofKind: "stripe";
        /**
         * @generated from protobuf field: wuxiaworld.api.v2.AddPaymentMethodRequest.AddStripePaymentMethod stripe = 1;
         */
        stripe: AddPaymentMethodRequest_AddStripePaymentMethod;
    } | {
        oneofKind: "braintree";
        /**
         * @generated from protobuf field: wuxiaworld.api.v2.AddPaymentMethodRequest.AddBraintreePaymentMethod braintree = 2;
         */
        braintree: AddPaymentMethodRequest_AddBraintreePaymentMethod;
    } | {
        oneofKind: undefined;
    };
}
/**
 * @generated from protobuf message wuxiaworld.api.v2.AddPaymentMethodRequest.AddStripePaymentMethod
 */
export interface AddPaymentMethodRequest_AddStripePaymentMethod {
    /**
     * @generated from protobuf field: google.protobuf.StringValue idempotencyKey = 1;
     */
    idempotencyKey?: StringValue;
}
/**
 * @generated from protobuf message wuxiaworld.api.v2.AddPaymentMethodRequest.AddBraintreePaymentMethod
 */
export interface AddPaymentMethodRequest_AddBraintreePaymentMethod {
    /**
     * @generated from protobuf field: google.protobuf.StringValue nonce = 1;
     */
    nonce?: StringValue;
    /**
     * @generated from protobuf field: google.protobuf.StringValue deviceData = 2;
     */
    deviceData?: StringValue;
}
// ...
export interface GetGatewayTokenRequest {
    /**
     * @generated from protobuf field: wuxiaworld.api.v2.PaymentMethodGateway gateway = 1;
     */
    gateway: PaymentMethodGateway;
    /**
     * @generated from protobuf field: wuxiaworld.api.v2.GetGatewayTokenRequest.GatewayTokenType tokenType = 2;
     */
    tokenType: GetGatewayTokenRequest_GatewayTokenType;
}
/**
 * @generated from protobuf enum wuxiaworld.api.v2.GetGatewayTokenRequest.GatewayTokenType
 */
export enum GetGatewayTokenRequest_GatewayTokenType {
    /**
     * @generated from protobuf enum value: Default = 0;
     */
    Default = 0,
    /**
     * @generated from protobuf enum value: AddPaymentMethod = 1;
     */
    AddPaymentMethod = 1,
    /**
     * @generated from protobuf enum value: Checkout = 2;
     */
    Checkout = 2
}
// ...
class AddPaymentMethodRequest$Type extends MessageType<AddPaymentMethodRequest> {
    constructor() {
        super("wuxiaworld.api.v2.AddPaymentMethodRequest", [
            { no: 1, name: "stripe", kind: "message", oneof: "paymentMethod", T: () => AddPaymentMethodRequest_AddStripePaymentMethod },
            { no: 2, name: "braintree", kind: "message", oneof: "paymentMethod", T: () => AddPaymentMethodRequest_AddBraintreePaymentMethod }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message wuxiaworld.api.v2.AddPaymentMethodRequest
 */
export const AddPaymentMethodRequest = new AddPaymentMethodRequest$Type();
// @generated message type with reflection information, may provide speed optimized methods
class AddPaymentMethodRequest_AddStripePaymentMethod$Type extends MessageType<AddPaymentMethodRequest_AddStripePaymentMethod> {
    constructor() {
        super("wuxiaworld.api.v2.AddPaymentMethodRequest.AddStripePaymentMethod", [
            { no: 1, name: "idempotencyKey", kind: "message", T: () => StringValue }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message wuxiaworld.api.v2.AddPaymentMethodRequest.AddStripePaymentMethod
 */
export const AddPaymentMethodRequest_AddStripePaymentMethod = new AddPaymentMethodRequest_AddStripePaymentMethod$Type();
// @generated message type with reflection information, may provide speed optimized methods
class AddPaymentMethodRequest_AddBraintreePaymentMethod$Type extends MessageType<AddPaymentMethodRequest_AddBraintreePaymentMethod> {
    constructor() {
        super("wuxiaworld.api.v2.AddPaymentMethodRequest.AddBraintreePaymentMethod", [
            { no: 1, name: "nonce", kind: "message", T: () => StringValue },
            { no: 2, name: "deviceData", kind: "message", T: () => StringValue }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message wuxiaworld.api.v2.AddPaymentMethodRequest.AddBraintreePaymentMethod
 */
export const AddPaymentMethodRequest_AddBraintreePaymentMethod = new AddPaymentMethodRequest_AddBraintreePaymentMethod$Type();
// ...
class GetGatewayTokenRequest$Type extends MessageType<GetGatewayTokenRequest> {
    constructor() {
        super("wuxiaworld.api.v2.GetGatewayTokenRequest", [
            { no: 1, name: "gateway", kind: "enum", T: () => ["wuxiaworld.api.v2.PaymentMethodGateway", PaymentMethodGateway] },
            { no: 2, name: "tokenType", kind: "enum", T: () => ["wuxiaworld.api.v2.GetGatewayTokenRequest.GatewayTokenType", GetGatewayTokenRequest_GatewayTokenType] }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message wuxiaworld.api.v2.GetGatewayTokenRequest
 */
export const GetGatewayTokenRequest = new GetGatewayTokenRequest$Type();